requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
bcrypt>=4.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
email-validator>=2.0.0
PyJWT>=2.8.0